since sdss are trained by autoencoder, we don't need to keep dataset separate; also fix saved galaxy test
